backend/requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
requests
google-auth
python-dotenv
reportlab
Pillow
